# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Tributary API",
        "description": "Multi-regime e-invoicing reference implementation (CO/ES/DE). Not certified under any fiscal regime \u2014 see ADR-005 and the README's scope-limits section. No login endpoint: this is a pure OAuth2 resource server, tokens are issued elsewhere.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:18081",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/v1/invoices": {
            "post": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Registered in DRAFT. Idempotent: re-registering the same businessKey returns the existing draft.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponseDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already registered in a different state"
                    },
                    "422": {
                        "description": "EN 16931 validation violations"
                    }
                }
            }
        },
        "/api/v1/invoices/{businessKey}/issuances": {
            "post": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "issue",
                "parameters": [
                    {
                        "name": "businessKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Issued; body carries the regime-resolved state",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponseDto"
                                }
                            }
                        }
                    },
                    "424": {
                        "description": "Regime unreachable \u2014 document left in NEEDS_RECONCILIATION, never retried blindly (ADR-003)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Invalid state for issuance"
                    }
                }
            }
        },
        "/api/v1/invoices/{businessKey}/corrections": {
            "post": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "correct",
                "parameters": [
                    {
                        "name": "businessKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CorrectionRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Correcting document created (RF-004: never an edit)"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Invalid state, or the regime refused the correction"
                    }
                }
            }
        },
        "/api/v1/records/{recordId}/verification": {
            "get": {
                "tags": [
                    "record-controller"
                ],
                "operationId": "verify",
                "parameters": [
                    {
                        "name": "recordId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecordVerificationView"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/invoices/{businessKey}": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "get",
                "parameters": [
                    {
                        "name": "businessKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No invoice with that businessKey"
                    }
                }
            }
        },
        "/api/v1/invoices/{businessKey}/renderings/xrechnung": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "renderXRechnung",
                "parameters": [
                    {
                        "name": "businessKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EN 16931 / XRechnung CII document",
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Mapping constraint violated \u2014 never a silently invalid document"
                    }
                }
            }
        },
        "/api/v1/chains/{chainId}/verification": {
            "get": {
                "tags": [
                    "chain-controller"
                ],
                "operationId": "verify_1",
                "parameters": [
                    {
                        "name": "chainId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "INTACT with recordsVerified, or BROKEN with brokenRecordId, storedHash, recomputedHash, totalMismatches and recordsVerified",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChainVerificationView"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown chain, or a chain with zero records \u2014 indistinguishable, so answered honestly as not found"
                    }
                }
            }
        },
        "/api/v1/subjects/{subjectId}/personal-data": {
            "delete": {
                "tags": [
                    "personal-data-controller"
                ],
                "operationId": "suppress",
                "parameters": [
                    {
                        "name": "subjectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonalDataSuppressionRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Idempotent: SUPPRESSED, or ALREADY_SUPPRESSED if the key was already destroyed (RF-007)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Any role other than ADMIN \u2014 issuance and evidence destruction never share an identity (CV-08)"
                    },
                    "409": {
                        "description": "An active fiscal retention obligation still covers this subject"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BuyerDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "taxIdentifier": {
                        "type": "string"
                    },
                    "countryCode": {
                        "type": "string",
                        "minLength": 1
                    },
                    "address": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "countryCode",
                    "name"
                ]
            },
            "InvoiceLineDto": {
                "type": "object",
                "properties": {
                    "lineIdentifier": {
                        "type": "string",
                        "minLength": 1
                    },
                    "itemName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "unitCode": {
                        "type": "string",
                        "minLength": 1
                    },
                    "unitPrice": {
                        "type": "number"
                    },
                    "lineDiscount": {
                        "type": "number"
                    },
                    "taxCategory": {
                        "type": "string",
                        "minLength": 1
                    },
                    "taxRate": {
                        "type": "number"
                    },
                    "vatExemptionReason": {
                        "type": "string"
                    }
                },
                "required": [
                    "itemName",
                    "lineIdentifier",
                    "quantity",
                    "taxCategory",
                    "taxRate",
                    "unitCode",
                    "unitPrice"
                ]
            },
            "InvoiceRequestDto": {
                "type": "object",
                "properties": {
                    "saleId": {
                        "type": "string",
                        "minLength": 1
                    },
                    "issuer": {
                        "$ref": "#/components/schemas/IssuerDto"
                    },
                    "buyer": {
                        "$ref": "#/components/schemas/BuyerDto"
                    },
                    "currency": {
                        "type": "string",
                        "minLength": 1
                    },
                    "issueDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceLineDto"
                        }
                    },
                    "documentLevelAllowance": {
                        "type": "number"
                    }
                },
                "required": [
                    "buyer",
                    "currency",
                    "issueDate",
                    "issuer",
                    "lines",
                    "saleId"
                ]
            },
            "IssuerDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "taxIdentifier": {
                        "type": "string",
                        "minLength": 1
                    },
                    "countryCode": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "countryCode",
                    "name",
                    "taxIdentifier"
                ]
            },
            "InvoiceResponseDto": {
                "type": "object",
                "properties": {
                    "businessKey": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "issuerTaxIdentifier": {
                        "type": "string"
                    },
                    "buyerName": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "issueDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "taxExclusiveAmount": {
                        "type": "number"
                    },
                    "taxTotal": {
                        "type": "number"
                    },
                    "taxInclusiveAmount": {
                        "type": "number"
                    }
                }
            },
            "CorrectionRequestDto": {
                "type": "object",
                "properties": {
                    "reason": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "reason"
                ]
            },
            "RecordVerificationView": {
                "type": "object",
                "properties": {
                    "recordId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "hash": {
                        "type": "string"
                    },
                    "previousHash": {
                        "type": "string"
                    },
                    "chainPosition": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "issuedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "nonSubmittedNotice": {
                        "type": "string"
                    }
                }
            },
            "ChainVerificationView": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "recordsVerified": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "brokenRecordId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "storedHash": {
                        "type": "string"
                    },
                    "recomputedHash": {
                        "type": "string"
                    },
                    "totalMismatches": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "PersonalDataSuppressionRequestDto": {
                "type": "object",
                "properties": {
                    "justification": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "justification"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}